requirements.txt
deap==1.4.1
gymnasium==1.0.0
matplotlib==3.9.3
numpy==2.1.3
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
torch==2.5.1
